feat(aep-api): wire the OpenChoreo runtime reader into the composition root
The coding-cycle watcher and both progress surfaces no longer depend on the
cluster gateway proxy, and a missing OBSERVER_URL is reported as the loss of
cycle log history rather than passing silently.
